Transfer make_trial_move to Ensemble
